chore(client): clean prototype
