+++ client/src/router.tsx
@@ -4,7 +4,6 @@ import { Home } from "./pages/Home";
 import { PingWs } from "./pages/PingWs";
 import { PingRest } from "./pages/PingRest";
 import { Coords } from "./pages/Coords";
-import { FirstIsolated } from "./isolated/v1/FirstIsolated";
 
 export const Router = (): JSX.Element => {
   return (
@@ -14,8 +13,6 @@ export const Router = (): JSX.Element => {
         <Route path="/ping-ws" element={<PingWs></PingWs>} />
         <Route path="/ping-rest" element={<PingRest></PingRest>} />
         <Route path="/coords" element={<Coords></Coords>} />
-        
-        <Route path="/isolated/v1" element={<FirstIsolated></FirstIsolated>} />
       </Routes>
     </BrowserRouter>
   );
